refactor(escrow): split escrow logic into a dedicated module
Move release_milestone/is_milestone_overdue, refund_unreleased_milestones/
cancel_contract, and raise_dispute/resolve_dispute out of the crate-root
lib.rs and into dedicated release.rs, refund.rs, and dispute.rs modules,
each contributing its own #[contractimpl] block to the Escrow contract
(the same pattern governance.rs already uses). lib.rs now owns only setup,
custody, reads, reputation, work evidence, and fee withdrawal.

Also removes contracts/escrow/src/refund_impl.rs, a pre-existing
alternate refund implementation that was never `mod`-declared (dead,
non-compiling code superseded by the real logic now in refund.rs).

Function bodies were moved verbatim; entrypoint signatures, error codes,
event topics, and storage keys are unchanged, so the public ABI is
identical. Verified with a byte-for-byte diff of per-test pass/fail
results between this branch and main (see PR description).
